product updated

+++ Modules/Product/Database/Migrations/2023_07_03_071531_create_products_table.php
@@ -36,7 +36,7 @@ return new class extends Migration
             $table->bigInteger('width')->default(0);
             $table->bigInteger('height')->default(0);
             $table->bigInteger('weight_with_packaging')->default(0);
-            $table->integer('unit_weight')->default(0);
+            $table->unsignedBigInteger('unit_weight')->default(0);
             $table->tinyInteger('status');
             $table->bigInteger('barcode')->nullable();
             $table->unsignedBigInteger('creator')->nullable();
